+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- BREAKING: Hide `Inode::set_blocks`
+- Add `Inode::fs_blocks` and `Inode::set_fs_blocks` for cleaner API that doesn't have to consider 512-byte vs full block sizes.
 - Disable writing when unsupported RO features are present to prevent corruption.
 - fsck fixes
     - Includes fixes that prevent data loss, generally allows better interop with other drivers due to checksum fixes